Змінено amount на integer у створенні ордерів
API може очікувати кількість токенів як ціле число (int)
замість числа з плаваючою точкою (float).

+++ predict_bot.py
@@ -458,7 +458,7 @@ class PredictFunBot:
             "tokenId": str(token_id),
             "side": side,
             "price": float(price),
-            "amount": float(amount)
+            "amount": int(amount)
         }
 
         if maker_address:
